requirements.txt
asyncio
aiohttp>=3.8.0
numpy>=1.21.0
pandas>=1.3.0
scikit-learn>=1.0.0
prometheus-client>=0.14.0
kubernetes>=24.0.0
docker>=6.0.0
pydantic>=1.10.0
fastapi>=0.85.0
uvicorn>=0.18.0
python-multipart>=0.0.5
jinja2>=3.0.0
python-dotenv>=0.20.0
PyYAML>=6.0
requests>=2.28.0
psutil>=5.9.0